+++ LinkTools/Identifiers.swift
@@ -18,4 +18,5 @@ enum Identifiers: String {
   case gui = "io.github.halo.LinkLiar"
   case helper = "io.github.halo.linkhelper"
   case daemon = "io.github.halo.linkdaemon"
+  case launcher = "io.github.halo.LinkLauncher"
 }